Rename TitleBarComponent
Make consistent with other component naming
