+++ src/comp/title-bar.component.ts
@@ -4,6 +4,7 @@ import {RouterLink} from 'angular2/router';
 @Component({
     selector: 'title-bar',
     directives: [RouterLink],
+    styles: ['nav { margin-bottom: 1em; }'],
     template: `
     <nav class="navbar navbar-dark bg-primary">
         <div class="nav navbar-nav">
